# contracts/v2/openapi-v2.json
{
  "components": {
    "schemas": {
...
      "ArtifactResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArtifactData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
...
      "CapabilitiesResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CapabilitiesData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
      "EvaluationResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EvaluationData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "ObservationResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ObservationData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "ObservationResponse",
        "type": "object"
      },
...
      "V2ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V2ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "error"
        ],
        "title": "V2ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Task-driven, evidence-native EO agent environment contract. Schema version 2.0.0.",
    "title": "EO Harness Environment API V2",
    "version": "0.4.0",
    "x-api-version": "v2",
    "x-schema-version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v2/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_v2_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "pattern": "^art-[a-f0-9]{64}$",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "get_artifact_content_v2_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "pattern": "^art-[a-f0-9]{64}$",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get Artifact Content",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/capabilities": {
      "get": {
        "operationId": "capabilities_v2_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Capabilities",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/evaluation": {
      "get": {
        "operationId": "get_evaluation_v2_episodes__episode_id__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/observations/{observation_id}": {
      "get": {
        "operationId": "get_observation_v2_episodes__episode_id__observations__observation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "pattern": "^obs-[a-f0-9]{32}$",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get Observation",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/replay": {
      "post": {
        "operationId": "replay_v2_episodes__episode_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Replay",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/state": {
      "get": {
        "operationId": "get_state_v2_episodes__episode_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get State",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/step": {
      "post": {
        "operationId": "step_v2_episodes__episode_id__step_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Step",
        "tags": [
          "V2 Environment"
        ]
      }
    },
...
  }
}